Decouple backend access from EthApi (#331)
* Decouple backend access from EthApi

* Module instead arbitrary struct

* Fix checker
